# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Discodeit API 문서",
    "description": "Discodeit 프로젝트의 Swagger API 문서입니다."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "로컬 서버"
    }
  ],
...
  "paths": {
...
    "/api/binaryContents": {
      "get": {
        "tags": [
          "BinaryContent"
        ],
        "summary": "여러 첨부 파일 조회",
        "operationId": "findAllByIdIn",
        "parameters": [
          {
            "name": "binaryContentIds",
            "in": "query",
            "description": "조회할 첨부 파일 ID 목록",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "첨부 파일 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BinaryContent"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/binaryContents/{binaryContentId}": {
      "get": {
        "tags": [
          "BinaryContent"
        ],
        "summary": "첨부 파일 조회",
        "operationId": "find",
        "parameters": [
          {
            "name": "binaryContentId",
            "in": "path",
            "description": "조회할 첨부 파일 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "첨부 파일 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BinaryContent"
                }
              }
            }
          },
          "404": {
            "description": "첨부 파일을 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "BinaryContent with id {binaryContentId} not found"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BinaryContent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "fileName": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "contentType": {
            "type": "string"
          },
          "bytes": {
            "type": "string",
            "format": "byte"
          }
        }
      }
    }
  }
}